Widen codegen scope IDs to avoid overflow on huge inputs
- change symbol scope_id storage from uint16_t to uint32_t
- change codegen scope_id and next_scope_id fields to uint32_t
- update overflow guard from UINT16_MAX to UINT32_MAX
- keep existing scope semantics while making exhaustion effectively
  unreachable for practical workloads

+++ codegen.c
@@ -329,7 +329,7 @@ void cg_code_begin(struct cg_state *s, bool in_function)
   code->varnames = object_new_list(arena);
   code->freevars = object_new_list(arena);
   code->cellvars = object_new_list(arena);
-  if (s->next_scope_id == UINT16_MAX) {
+  if (s->next_scope_id == UINT32_MAX) {
     internal_error("Ran out of scope identifiers");
   }
   code->scope_id = s->next_scope_id++;
